refactor auth-controller

+++ entity/user.go
@@ -3,7 +3,7 @@ package entity
 import "github.com/rocketlaunchr/dbq/v2"
 
 type User struct {
-	Id   int64  `json:"id" dbq:"user_id"`
+	Id   int    `json:"id" dbq:"user_id"`
 	Name string `json:"name" dbq:"user_name"`
 }
